Moved the logic of coercion from `movel` to `coercing`, to avoid code duplication and to improve the logic of coercion from String to DS in EVAL.
